añadir sonidos nuevos UI, mejorar interfaces de menus, folder, casos y minijuegos, configuracion Avatar lista para ser Camilo o Camila, puntos profesionales y sociales por interacciones sociales, escenas de decision por cantidad de puntos, Load game, interaccion entre menus, menu pausa

+++ Assets/MAIN/Scripts/Minigames/Find the error/Loaders/DocumentBaseJson.cs
@@ -33,7 +33,7 @@ namespace EXERCISE.Loaders
     [Serializable]
     public class RegionsJson
     {
-        public float pointsPerCorrect = 1f;
+        public int pointsPerCorrect = 1;
         public List<RegionJson> regions;
     }
 
@@ -43,7 +43,7 @@ namespace EXERCISE.Loaders
         public string id;
         public bool enabled;
         public bool correct;
-        public float points;
+        public int points;
         public List<SpanJson> spans;
     }
